breakdown orchestratror, clean code
