+++ docs/outdated_architecture/S04.3_Autonomy & Heartbeat (The `Idle` State).md
@@ -8,7 +8,7 @@
 2. Architecture & Design Rules
     
 
-A. The Heartbeat Timer (`src/orchestrator/heartbeat.rs`)
+A. The Heartbeat Timer (`src/orchestrator/heartbeat/monitor.rs`, `spawn_heartbeat_monitor`; `heartbeat/mod.rs` is export-only)
 
 A lightweight `tokio` background task monitors the timestamp of the last user `stdin` event.
