+++ docs/updated_architecture/02_ORCHESTRATOR_LAYER.md
@@ -1,6 +1,8 @@
 # Orchestrator layer
 
-## Core type (`orchestrator/core.rs`)
+## Core type (`orchestrator/core/`)
+
+`orchestrator/core/mod.rs` only declares submodules and re-exports; the type and constructor live in `core/orchestrator.rs`, with `step`, tool dispatch, transitions, etc. in sibling files under `core/`.
 
 `Orchestrator<E: LlmEngine>` owns:
 
@@ -32,7 +34,7 @@ Rough pipeline:
      - Non-empty → tools on + matched names for JIT guidance / possible targeted schema subset path.
 5. **Inner loop** (tool rounds, recovery): assemble system prompt (conversational vs full tools vs selected tools), optional JIT descriptor block, `build_llm_view`, `engine.generate`.
 6. **Interrupt:** `tokio::select!` on `interrupt_rx.changed()` → clear stack, inject idle/agenda prompt, return `FcpError::Interrupted`.
-7. **Parse** assistant JSON → `LlmResponse`; push to stack; condensation if over token threshold (`context_window`); interpret `LoopDirective`; execute tools through gatekeeper; apply `orchestrator::loop::*` policies.
+7. **Parse** assistant JSON → `LlmResponse`; push to stack; condensation if over token threshold (`orchestrator::context`, implemented in `context/window.rs`); interpret `LoopDirective`; execute tools through gatekeeper; apply `orchestrator::loop::*` policies.
 
 ### Assistant text on the TUI (`emit_optional_user_message`)
 
@@ -50,21 +52,20 @@ When there are **no** tool calls, behavior is unchanged: non-empty `message_to_u
 - **`normalize_tool_calls`:** e.g. `action` → `name`, agenda `id` → `task_id` in args.
 - **`LoopDirective`:** what the coordinator does next (execute tools, halt, recover, shift reflection).
 
-## Context assembly (`orchestrator/context.rs`)
+## Context pipeline (`orchestrator/context/`)
 
-- **`assemble`:** identity + **all** allowed tools for current `AgentState` as OpenAI-style JSON functions between `FCP_TOOL_DEFS_BEGIN/END` markers.
-- **`assemble_with_selected_tools`:** filters to a subset (used when targeting after routing).
-- **`assemble_conversational`:** JSON schema rules but **no** tool definitions (tool_calls expected empty).
+Submodules (all re-exported at `crate::orchestrator::context::…`):
+
+- **`assembler.rs` — `ContextAssembler`:** builds system prompts from identity + tool JSON.
+  - **`assemble`:** identity + **all** allowed tools for current `AgentState` as OpenAI-style JSON functions between `FCP_TOOL_DEFS_BEGIN/END` markers (`view` module).
+  - **`assemble_with_selected_tools`:** filters to a subset (used when targeting after routing).
+  - **`assemble_slim_tool_map`:** phrase map from **`compendium`** + tool defs without `parameters`.
+  - **`assemble_conversational`:** JSON schema rules but **no** tool definitions (tool_calls expected empty).
+- **`view.rs`:** `build_llm_view`, `ContextViewSettings`, delimiters. Clones `chat_stack` and may strip/slim tool `parameters`, trim tool result snippets per `ToolContextViewHint`, compact assistant JSON when configured.
+- **`window.rs`:** sliding-window condensation, rolling summary JSON, `plan_sliding_condensation`, etc.
+- **`compendium.rs`:** `build_phrase_compendium` / typical phrasing lines for slim tool prompts.
 
 System prompt text mandates **single JSON object** output, no markdown fences; model uses `FormatType::Json` in Ollama.
-
-## LLM-only view (`orchestrator/context_view.rs`)
-
-`build_llm_view` clones `chat_stack` and may:
-
-- Strip or slim tool `parameters` in the tool-def block (when `optimize_context` + not recovery pass).
-- Trim tool result snippets per `ToolContextViewHint` and config overrides.
-- Compact assistant JSON when configured.
 
 `Orchestrator::force_full_tool_schemas_in_llm_view` forces full schemas for one pass after certain gatekeeper schema faults.
 
@@ -76,22 +77,20 @@ System prompt text mandates **single JSON object** output, no markdown fences; m
 
 Note: **Pre-LLM** routing uses **user input** string, not the model’s `thought` field (despite the parameter name in `match_tools`).
 
-## Heartbeat (`orchestrator/heartbeat.rs`)
+## Heartbeat (`orchestrator/heartbeat/`)
 
-Polls every second: if `now - last_input_time > idle_timeout_secs`, sends on `watch` channel. Orchestrator uses this to enter idle injection (and clears stack in that branch).
+`heartbeat/monitor.rs` (`spawn_heartbeat_monitor`): polls every second; if `now - last_input_time > idle_timeout_secs`, sends on `watch` channel. Orchestrator uses this to enter idle injection (and clears stack in that branch).
 
-## Alarms (`orchestrator/alarm_scheduler.rs`, `tools/clock`)
+## Alarms (`orchestrator/alarms/`, `tools/clock`)
 
-- `alarms.json` stores `AlarmRecord` rows with `fire_at_unix`; scheduler sleeps until next due, fires `TuiEvent::SystemAlarm`, persists remaining rows.
+- **`alarms/scheduler.rs`:** `spawn_alarm_scheduler` — `alarms.json` stores `AlarmRecord` rows with `fire_at_unix`; sleeps until next due, fires `TuiEvent::SystemAlarm`, persists remaining rows.
 - Agenda-linked alarms include `agenda_task_id` → UI sends `UserAction::AgendaAlarmPending` with confirmation block.
+- **`alarms/missed_startup.rs`:** `startup_overdue_agenda_hint` — startup banner if agenda-linked alarms were due while the app was offline.
 
-## Missed agenda (`orchestrator/missed_agenda.rs`)
+## LLM support text (`orchestrator/llm_support/`)
 
-Startup hint if overdue tasks exist (spawned once).
-
-## Post-tool guidance (`orchestrator/post_tool_guidance.rs`)
-
-Injected strings for recovery / user-facing follow-ups after tool success or failure.
+- **`llm_support/json_envelope.rs`:** split leading JSON object from trailing prose, trailing-content protocol check, `llm_json_parse_recovery_message` for `RecoverFromFuckup`.
+- **`llm_support/post_tool_guidance.rs`:** `POST_TOOL_*` strings and `recover_override_message_for_tool_failure` injected after tool batches / failures.
 
 ## Loop policy modules (`orchestrator/loop/`)
 
@@ -109,4 +108,4 @@ stateDiagram-v2
     Idle --> Chat: new user input
 ```
 
-(Exact transitions are implemented in `core.rs` + policies; this diagram is illustrative.)
+(Exact transitions are implemented in `orchestrator/core/` + `orchestrator/loop/` policies; this diagram is illustrative.)
