+++ src/orchestrator/context/compendium.rs
