+++ src/orchestrator/core/deck.rs
@@ -1,0 +1,133 @@
+use crate::engine::LlmEngine;
+use crate::orchestrator::llm_support::json_envelope::split_leading_json_object;
+use crate::orchestrator::state::LlmResponse;
+
+use super::Orchestrator;
+
+impl<E: LlmEngine> Orchestrator<E> {
+    /// Emits an assistant-facing message to TUI when present in the model JSON.
+    /// Tool rounds: only `message_to_user` is shown on the main transcript; tool names go to Status
+    /// (and full payloads remain in tracing via existing LLM/tool logs).
+    pub(super) async fn emit_optional_user_message(&mut self, response_content: &str) {
+        let Some(tx) = &self.tui_tx else {
+            return;
+        };
+
+        let json_slice = split_leading_json_object(response_content).0;
+        let Ok(parsed) = serde_json::from_str::<LlmResponse>(json_slice) else {
+            return;
+        };
+
+        let has_tools = !parsed.tool_calls.is_empty();
+        let msg_opt = parsed
+            .message_to_user
+            .as_ref()
+            .map(|s| s.trim())
+            .filter(|s| !s.is_empty())
+            .map(|s| s.to_string());
+
+        if has_tools {
+            let joined = parsed
+                .tool_calls
+                .iter()
+                .map(|t| t.name.as_str())
+                .collect::<Vec<_>>()
+                .join(", ");
+            tracing::debug!(
+                event = "UI_TOOL_ROUND_STATUS",
+                tool_count = parsed.tool_calls.len(),
+                tool_names = %joined,
+                "Tool round: names on status line; deck is message_to_user only"
+            );
+            let tool_names_line = format!("Tools: {}", Self::trim_chars(&joined, 88));
+            if let Some(msg) = msg_opt {
+                if self
+                    .last_deck_message_body
+                    .as_deref()
+                    .is_some_and(|prev| prev == msg.as_str())
+                {
+                    tracing::debug!(
+                        event = "UI_SKIP_DUPLICATE_DECK_MESSAGE",
+                        msg_len = msg.len(),
+                        preview = %msg.chars().take(120).collect::<String>(),
+                        "Skipping duplicate assistant deck message (tool round)"
+                    );
+                } else {
+                    self.last_deck_message_body = Some(msg.clone());
+                    let agent_name = self.agent_name();
+                    tracing::info!(
+                        event = "UI_EMIT_INCOMING_MESSAGE",
+                        agent = %agent_name,
+                        msg_len = msg.len(),
+                        preview = %msg.chars().take(120).collect::<String>(),
+                        "Emitting assistant message to TUI deck (tool round)"
+                    );
+                    let _ = tx
+                        .send(crate::ui::events::TuiEvent::IncomingMessage(format!(
+                            "[{}]: {}",
+                            agent_name, msg
+                        )))
+                        .await;
+                }
+            }
+            self.activity_line = Some(tool_names_line);
+            self.broadcast_state().await;
+            return;
+        }
+
+        let Some(msg) = msg_opt else {
+            self.activity_line = None;
+            self.broadcast_state().await;
+            return;
+        };
+
+        self.activity_line = None;
+        if self
+            .last_deck_message_body
+            .as_deref()
+            .is_some_and(|prev| prev == msg.as_str())
+        {
+            tracing::debug!(
+                event = "UI_SKIP_DUPLICATE_DECK_MESSAGE",
+                msg_len = msg.len(),
+                preview = %msg.chars().take(120).collect::<String>(),
+                "Skipping duplicate assistant deck message (same body as previous emit this step)"
+            );
+            self.broadcast_state().await;
+            return;
+        }
+        self.last_deck_message_body = Some(msg.clone());
+        let agent_name = self.agent_name();
+        tracing::info!(
+            event = "UI_EMIT_INCOMING_MESSAGE",
+            agent = %agent_name,
+            msg_len = msg.len(),
+            preview = %msg.chars().take(120).collect::<String>(),
+            "Emitting assistant message to TUI deck"
+        );
+        let _ = tx
+            .send(crate::ui::events::TuiEvent::IncomingMessage(format!(
+                "[{}]: {}",
+                agent_name, msg
+            )))
+            .await;
+        self.broadcast_state().await;
+    }
+
+    /// Deck line for cap-recovery when the normal JSON → deck path did not apply.
+    pub(super) async fn emit_assistant_deck_line(&mut self, msg: &str) {
+        let Some(tx) = &self.tui_tx else {
+            return;
+        };
+        let agent_name = self.agent_name();
+        let _ = tx
+            .send(crate::ui::events::TuiEvent::IncomingMessage(format!(
+                "[{}]: {}",
+                agent_name, msg
+            )))
+            .await;
+        self.last_deck_message_body = Some(msg.to_string());
+        self.activity_line = None;
+        self.broadcast_state().await;
+    }
+}
