+++ src/orchestrator/mod.rs
@@ -1,15 +1,10 @@
 pub mod state;
-pub mod json_envelope;
-pub mod post_tool_guidance;
+pub mod llm_support;
 pub mod context;
-pub mod context_view;
-pub mod context_window;
 pub mod heartbeat;
-pub mod alarm_scheduler;
-pub mod missed_agenda;
+pub mod alarms;
 pub mod core;
 pub mod tool_router;
-pub mod tool_compendium;
 pub mod r#loop;
 
 pub use self::core::Orchestrator;
