+++ src/tools/routing_phrases.rs
@@ -7,15 +7,15 @@ pub fn fallback_triggers(tool_name: &str) -> &'static str {
         "vault:read" => "reading files, checking notes, looking at documents, show me, what is in my vault, review notes, open file, read my notes",
         "vault:write" => "writing files, saving notes, creating documents, write this down, save to vault, take a note, jot down, record",
         "vault:list" => "listing files, what files do I have, show directory, browse vault, what is in my folder, list notes",
-        "memory:query" => "remembering, recalling, do you remember, what did I say, past conversations, search memory, who am I, what is my name, user name, my identity, preferences, facts about the user, recall, recognize me, history",
+        "memory:query" => "remembering, recalling, do you remember, what did I say, past conversations, search memory, who am I, what is my name, user name, my identity, preferences, facts about the user, recall, recognize me, history, look up contact, stored email address, email in memory, before sending email, before I mail, find their email, multi-step reminder, after the alarm, synthesis notes, 30_Synthesis, long-term recall, vector memory",
         "memory:commit" => "save staged entry to vault by staged_id, persist to disk when user asked to keep forever or save permanently",
         "memory:commit_all" => "flush all staged memories, persist all staged entries, bulk commit staged memory",
         "memory:staged_list" => "show staged memory ids, list staged entries, what is currently staged before commit",
         "memory:stage" => "ephemeral staging with ttl, no vault write until commit, hold fact until user wants disk save",
-        "agenda:push" => "adding tasks, to-do list, add to agenda queue, schedule, plan, new task without setting a time",
+        "agenda:push" => "adding tasks, to-do list, add to agenda queue, schedule, plan, new task without setting a time, queue work, background task, add to my list, track this for later",
         "agenda:list" => "show tasks, what is on my list, pending items, show agenda, my schedule, what do I have to do",
         "agenda:remove" => "remove task, cancel agenda item, delete from list, drop task, never mind that reminder, scratch that task",
-        "agenda:remind_at" => "remind me at, remind me in, remind me about, remind me tomorrow, remember to, nudge me at, ping me at, todo reminder, snooze task, alarm for my task, at 3pm for this, on my agenda, on my todo list, task_id reminder",
+        "agenda:remind_at" => "remind me at, remind me in, remind me about, remind me tomorrow, remember to, nudge me at, ping me at, todo reminder, snooze task, alarm for my task, at 3pm for this, on my agenda, on my todo list, task_id reminder, in two minutes, in 2 minutes, multi-step task, several steps later, then send email, remind yourself to, delayed checklist",
         "agenda:complete" => "finishing tasks, mark done, complete task, check off, task finished, I did it",
         "web:fetch" => "fetching URLs, web search, look up online, check website, browse internet, search the web, what is happening, news, look this up",
         "web:artifact_query" => "query fetched web artifact by artifact id, search fetched page snippets, retrieve specific sections from buffered webpage",
@@ -26,9 +26,9 @@ pub fn fallback_triggers(tool_name: &str) -> &'static str {
         "weather:current" => "weather now, temperature outside, is it raining, rainfall, sunny or cloudy, conditions today, current conditions, what's the weather like",
         "weather:forecast" => "weather forecast, hourly temperature, next days weather, will it rain tomorrow, rain outlook, sun or clouds, upcoming weather",
         "wiki:summary" => "wikipedia, encyclopedia, what is X, who was, summary of topic, general knowledge, what does wikipedia say, define concept, historical figure, science topic overview",
-        "mail:check" => "check email, new mail, inbox, unread messages, check gmail, any new emails, email summary, recent emails, list messages",
-        "mail:read" => "read email, open message, show email, email details, message content, full email, read message, opens mail",
-        "mail:write" => "send email, compose mail, write email, reply, email to, send a message, draft email",
+        "mail:check" => "check email, new mail, inbox, unread messages, check gmail, any new emails, email summary, recent emails, list messages, gmail search, list threads, search inbox, from filter, mail listing",
+        "mail:read" => "read email, open message, show email, email details, message content, full email, read message, opens mail, full gmail body, message id, open thread body",
+        "mail:write" => "send email, compose mail, write email, reply, email to, send a message, draft email, dispatch mail, send the greeting",
         "mail:digest" => "summarize email, today's mail, digest inbox, what came in today, recap messages, overview of recent gmail, batch list mail",
         "mail:delete" => "delete email, trash message, remove email, get rid of mail, discard message",
         "mail:move" => "move email to folder, label message, file under label, move to spam, archive to label, put mail in ebay folder",
